create core package and move errors over there
