+++ package.json
@@ -67,6 +67,7 @@
     "typescript": "4.9.5"
   },
   "workspaces": [
+    "packages/core",
     "packages/runtype",
     "packages/router",
     "packages/http",
