+++ packages/cli/package.json
@@ -32,6 +32,7 @@
     "url": "https://github.com/MionKit/mion/issues"
   },
   "dependencies": {
+    "@mionkit/core": "^0.4.2",
     "commander": "^8.3.0"
   }
 }
