+++ packages/client/package.json
@@ -55,6 +55,7 @@
     "url": "https://github.com/MionKit/mion/issues"
   },
   "dependencies": {
+    "@mionkit/core": "^0.4.2",
     "@deepkit/type": "^1.0.1-alpha.97",
     "@mionkit/router": "^0.4.2"
   },
