+++ packages/client/tsconfig.json
@@ -6,5 +6,5 @@
   "reflection": true,
   "include": ["."],
   "exclude": ["examples", ".dist"],
-  "references": [{"path": "../router"}]
+  "references": [{"path": "../core"}, {"path": "../router"}]
 }
