+++ packages/core/src/constants.ts
@@ -1,0 +1,13 @@
+/* ########
+ * 2022 mion
+ * Author: Ma-jerez
+ * License: MIT
+ * The software is provided "as is", without warranty of any kind.
+ * ######## */
+
+import {CoreOptions} from './types';
+
+export const DEFAULT_CORE_OPTIONS: CoreOptions = {
+    /** automatically generate and uuid */
+    autoGenerateErrorId: false,
+};
