+++ packages/core/src/errors.spec.ts
@@ -5,20 +5,17 @@
  * The software is provided "as is", without warranty of any kind.
  * ######## */
 
-import {RouteError} from '@mionkit/core';
-import {initRouter, resetRouter} from './router';
+import {RouteError, setErrorOptions} from './routeError';
 
 describe('Route errors should', () => {
     it('automatically generate an dis when RouteOptions autoGenerateErrorId is set to true', () => {
-        initRouter({autoGenerateErrorId: true});
+        setErrorOptions({autoGenerateErrorId: true});
         const error = new RouteError({statusCode: 400, publicMessage: 'error'});
 
         expect(typeof error.id).toEqual('string');
         expect((error.id as string).length).toEqual(61);
 
-        resetRouter();
-
-        initRouter({autoGenerateErrorId: false});
+        setErrorOptions({autoGenerateErrorId: false});
         const error2 = new RouteError({statusCode: 400, publicMessage: 'error'});
         expect(error2.id).toEqual(undefined);
     });
