+++ packages/http/package.json
@@ -55,6 +55,7 @@
     "url": "https://github.com/MionKit/mion/issues"
   },
   "dependencies": {
+    "@mionkit/core": "^0.4.2",
     "@mionkit/router": "^0.4.2"
   },
   "gitHead": "2efdcd77c572e64f2b9621b4921ffd2f5e83bc0a"
