+++ packages/http/tsconfig.build.json
@@ -6,5 +6,5 @@
   "reflection": false,
   "include": ["."],
   "exclude": ["node_modules", "**/*.spec.ts", "examples", ".dist"],
-  "references": [{"path": "../router"}]
+  "references": [{"path": "../core"}, {"path": "../router"}]
 }
