+++ packages/router/tsconfig.json
@@ -5,6 +5,6 @@
   },
   "include": ["."],
   "exclude": ["examples", ".dist"],
-  "references": [{"path": "../runtype"}],
+  "references": [{"path": "../core"}, {"path": "../runtype"}],
   "reflection": true
 }
