+++ tsconfig.json
@@ -22,6 +22,8 @@
   },
   "exclude": ["node_modules", ".dist"],
   "references": [
+    {"path": "packages/core"},
+    {"path": "packages/core"},
     {"path": "packages/runtype"},
     {"path": "packages/router"},
     {"path": "packages/http"},
